database: Move the readonly database into a separate module.
